Share Razor semantic tokens registration
